move component-layout out of /editing
