+++ packages/content/component-layout.d.ts
@@ -1,0 +1,1 @@
+﻿export * from './src/component-layout/index';
