+++ packages/content/src/component-layout/resolver.ts
